# content/src/main/resources/openapi-content.yml
openapi: 3.0.3
info:
  title: "Content Module API"
  version: 1.0.0
  description: "콘텐츠 모듈 API 명세"

paths:
  /api-v1/courses:
    post:
      tags:
        - Course
      summary: 새 강의 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateRequest'
      responses:
        '200':
          description: 강의 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Course
      summary: 강의 목록 조회
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: 정렬 기준 필드명
          schema:
            type: string
          example: "createdAt"
        - name: dir
          in: query
          description: 정렬 방향
          schema:
            $ref: '#/components/schemas/SortDirection'
          example: "ASC"
      responses:
        '200':
          description: 강의 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoursePageResponse'

  /api-v1/courses/{courseId}:
    get:
      tags:
        - Course
      summary: 강의 상세 조회
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 강의 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '404':
          description: 강의를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    # ==================== Request ====================
    CourseCreateRequest:
      type: object
      required:
        - title
        - description
        - instructorId
        - level
      properties:
        title:
          type: string
          example: "테스트 강의"
        description:
          type: string
          example: "테스트용 코스 설명"
        instructorId:
          type: string
          example: "255fdb4d-fdab-4d0b-95ba-420107b361bd"
        thumbnailUrl:
          type: string
          example: "https://example.com/thumbnail.jpg"
        level:
          type: string
          enum: [JUNIOR, MIDDLE, SENIOR, EXPERT]
          example: "JUNIOR"
        tags:
          type: array
          items:
            type: integer
          example: [1, 2, 3, 4]
        sections:
          type: array
          items:
            $ref: '#/components/schemas/SectionCreateRequest'

    SectionCreateRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          example: "섹션 1"
        lectures:
          type: array
          items:
            $ref: '#/components/schemas/LectureCreateRequest'

    LectureCreateRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          example: "강의 1"
        videoUrl:
          type: string
          example: "https://video.url/1"

    # ==================== Response ====================
    CourseResponse:
      type: object
      properties:
        id:
          type: string
          example: "479365be-0d40-4c7b-a6c8-ecc290d856fd"
        title:
          type: string
          example: "테스트 강의"
        description:
          type: string
          example: "테스트용 코스 설명"
        instructor:
          $ref: '#/components/schemas/InstructorResponse'
        level:
          $ref: '#/components/schemas/EnumResponse'
        durationInHours:
          type: integer
          example: 40
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'
        sections:
          type: array
          items:
            $ref: '#/components/schemas/SectionResponse'
        thumbnailUrl:
          type: string
        createdAt:
          type: string
          nullable: true
        updatedAt:
          type: string
          nullable: true

    CourseSummaryResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        instructor:
          $ref: '#/components/schemas/InstructorResponse'
        level:
          $ref: '#/components/schemas/EnumResponse'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagResponse'
        thumbnailUrl:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    CoursePageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CourseSummaryResponse'
        pageNumber:
          type: integer
        pageSize:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
        empty:
          type: boolean
        numberOfElements:
          type: integer

    SectionResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        durationInSeconds:
          type: integer
        order:
          type: integer
        lectures:
          type: array
          items:
            $ref: '#/components/schemas/LectureResponse'

    LectureResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        videoUrl:
          type: string
        order:
          type: integer
        durationInSeconds:
          type: integer

    InstructorResponse:
      type: object
      properties:
        instructorId:
          type: string
        name:
          type: string

    TagResponse:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        color:
          type: string
        variant:
          type: string

    EnumResponse:
      type: object
      properties:
        key:
          type: string
        value:
          type: string

    # ==================== Sort ====================
    SortDirection:
      type: string
      enum: [ASC, DESC]
      example: "DESC"

    # ==================== Error ====================
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        group:
          type: string
        timestamp:
          type: string
          format: date-time